Improves frontend tensors and traces
Removes the notion of `TraceLayer`. Instead, each frontend op now maintains
its own inputs and outputs (in the form of `TraceTensor`s) as we build the frontend
representation (consisting of frontend `Tensor`s). Creating the final trace is then
just a matter of walking backwards over the tensors and retrieving their `.op`s, which
are now effectively trace layers.