Adds developer documentation into the rendered documentation
Adds the guides from `docs/development` into the rendered Sphinx documentation.
This requires some changes to ensure that Myst is able to parse the files correctly.
One of the consequences of this is that the links in the markdown files under `docs/`
don't work except through Myst. Ideally, any files under `docs/` should not be looked at
directly but instead only via the rendered documentation.