Adds `build` method for FlatIR ops
The FlatIR ops previously had to manually define constructor since we relied
on the constructor to set up the producer of the output tensors.

This commit introduces a new `build` class method which will do that for us.
That means we are now able to remove all the manually defined constructors and
can rely on the constructors generated by `@dataclass`, as we do for trace operations.

Also removes the `source_op` field from FlatIR operations since we can determine
the source operation based on the tensors.