[mlir-tensorrt] migrate internal changes
This change is a compilation of the following:

[executor] Move some integrations tests under the executor project

This change moves some end-to-end tests under the Executor project. Any
integration test that can purely use the dialects and passes defined by
Executor or upstream MLIR should be located under the Executor project.
This also fixes Executor test configuration for release builds with
assertions off.

[executor/runtime] Fix debug-mode test failure due to runtime type check

We use function '__check_for_function' in the core runtime module to verify the
existence of builtin functions when loading the Lua program. Either the
introduction of this function or a change in compilation settings broke the
Debug build since it causes the lua metaprogramming library (sol2) to check
whether the loaded lua object is a bool (which is not true since we are loading
a function, not a bool). Fix the logic by explicitly loading a function and
verifying that the function is valid.

[StablehloExt] Add pass to lower special custom call operations

Adds a pass to lower `stablehlo.custom_call` that are produced by
frontends like JAX to handle special kinds of operations that map to
CHLO ops or sharding constraints.

[Dialect/StableHloExt] Avoid folding if number of elements  > kFoldOpEltLimit

This MR adds condition to avoid folding if number of elements > kFoldOpEltLimit
for constant folders where it was missing.

NFC: unify handling of ASAN flag for tests

[Dialect/TensorRT] Move the patterns for composing reassociate reshapes downstream

Upstream MLIR is changing the format for `tensor.collapse_shape`, which makes the
reassociative reshape template patterns that we utilize incompatible for use with
`tensorrt.expand_rank` and `tensorrt.collapse_rank`. This change pulls those
patterns down into dedicated classes in the TensorRT dialect code.

[Dialect/StableHloExt] Simplify stablehlo.broadcast_in_dim removing unnecessary rank expansion.

Introduces pattern to simplify stablehlo sequences of reshape->broadcast->reshape.

For example, it simplifies the following sequence

```
  %0 = stablehlo.reshape %arg0 : (tensor<1x1x1x256xf16>) -> tensor<1x1x1x1x1x1x1x1x256xf16>
  %1 = stablehlo.broadcast_in_dim %0, dims = [0, 1, 2, 3, 4, 5, 7, 8, 9] : (tensor<1x1x1x1x1x1x1x1x256xf16>) -> tensor<1x1x1x1x1x1x8x1x1x256x
f16>
  %2 = stablehlo.reshape %1 : (tensor<1x1x1x1x1x1x8x1x1x256xf16>) -> tensor<1x1x8x256xf16>
```

to

```
  %0 = stablehlo.broadcast_in_dim %arg0, dims = [0, 1, 2, 3] : (tensor<1x1x1x256xf16>) -> tensor<1x1x8x256xf16>
```

Co-authored-by: Sagar Shelke <sagshelke@nvidia.com>
Co-authored-by: Jinkyu Koo <jinkyuk@nvidia.com>