[Tripy][Docs] Recommend the use of `--pull always` to update the image in the contributing doc (#151)
`docker run` by default will only pull an image if it is entirely
missing, so the command given in `CONTRIBUTING.md` will not update the
image if it had been pulled before. This change adds the `--pull always`
setting in the example command so that it would check for an update
before running.