[Tripy][Bugfix] Ensure shape `__eq__` works with different lengths (#165)
This PR corrects a small bug with the `__eq__` implementation for
`tp.Shape`: The comparison was checking the `len` of the `shape` field
of the shape, but it should actually be checking the length of the
`tp.Shape` itself.

Note: The test case that was included in the unit tests worked "by
accident" because the shape in it was length 1, which is broadcasted up
to other shapes' lengths. Without this change, the test would fail if
comparing two shapes of different lengths where neither is length 1.