Updates the compiler to emit errors if evaluating while tracing
Updates the `tp.compile` API so that an error is emitted if a tensor
is evaluated while tracing for the purposes of compiling. Note that this
only applies to tensors that are part of the computation graph. If a tensor
is evaluated but not consumed by other tensors in the graph, we allow it.

For example, this is not allowed:
```py
def func(a):
    c = a + 1
    print(c)
    return c
```
whereas this is:
```py
def func(a):
    c = a + 1
    print(c.shape)
    return c
```