Removes `Shape` tensors
In Tripy, we want shapes of tensors to behave like lists. This makes it
easier and more natural to manipulate shapes. However, we previously achieved
this through a `Shape` abstraction that was actually a `Tensor` which behaved
like a list in some, but not all, ways, which made the API very confusing.

To solve this, we remove the `Shape` abstraction and instead represent
shapes of tensors as lists of `DimensionSize`s (which are similar to `ShapeScalar`s).
These lists can be manipulated freely and are then converted back to tensors
by the `convert_to_tensors` decorator, which takes over the functionality of
`convert_shape_inputs`. This also has the advantage that operations on shapes
are not materialized in the computation graph - only operations on the individual
dimensions are.

Also includes various miscellaneous fixes:
- Updates `expand` to disallow `-1`s in prepended dimensions.
- Updates `resize` to use function overloading to accept either output shape or scales.
- Simplifies logic in `flatten`, `repeat`, `unsqueeze`, `full`, and `pad`.
- Greatly simplifies logic in `reshape`.
- Disallows empty lists in `concatenate`, similar to what `stack` does.
- Moves `squeeze` into its own file.
- Fixes to `get_arg_candidate_column_offsets` to work with `skip_num_stack_entries`.
- Updates `Storage` op to copy list inputs so that changes made to a list after
    it is used to construct a tensor don't affect the created tensor.