[Runtime] Remove incorrect removal of previous entry in map in AllocT… (#324)
…racker::track; track DLPackTensors in RuntimeClient such that deleters
can be reset when RuntimeClient is destroyed.

This PR fixes two issues:
1. When you have a dlpack array allocated outside of `mlir-tensorrt`
(ex: numpy) and you create memref view of this array, and then create
another dlpack tensor from this memref, there was a bug in our existing
implementation of `AllocTracker::track`. In `track()` we were
unnecessarily reseting the value of already defined entry in map and
thereby resting the `externalReferenceCount`.
2. After fixing ^, and adding the test in `test_create_memref.py` it was
found that the `AllocTracker` (from `RuntimeClient) can be destructed
before the external allocated tensor is destructed. To fix this issue we
now remove the `deleter` of the tracked `dlpack` tensors when
`RuntimeClient` is destructed.

Thanks to @jhalakpatel for debugging this issue.