Enables execution of shell blocks during documentation generation
Updates our documentation generation logic to allow execution of shell
blocks in guides and READMEs. This allows us to ensure that the content
of shell blocks is always functional.

Consequently, we also remove the packaging test that was previously used
as a proxy for running the packaging instructions in the README.