Adds a new workflow to update package index
When we want to update to a new release of MLIR-TRT, we first need to publish
`packages.html` so we are able to `pip install` correctly. Previously, we were
doing this in each nightly, but it's better to do so whenever we push a change
to `packages.html` on `main`. That's what this workflow does.