[mlir-tensorrt][CI] Fix CPM cache path
This PR fixes CPM cache path in mlir-tensorrt CI.
GitHub actions has the following rule with cache.
'Workflow runs can restore caches created in either
the current branch or the default branch
(usually main)'.
Because of this cache was not being shared between
PRs. Also we did not run CI on main branch, thus
there is no cache saved for main branch.
We now make change to run LIT tests on main branch
so that cache is available to PR feature
branches.