Integrate internal changes (#374)
This is a combination of the following commits:

## [compiler] Drop the StableHLO "signature refinement" API

Previously, we added a special API for TriPy to be able to perform only
pre-processing (including entrypoint signature type refinement) on the
input module. This path is no longer needed by TriPy, and since there
are no
other customers, it can be safely dropped from the C/C++/Python APIs.

## [executor/runtime] Update NCCL module to enable non-blocking
communicator

Updates the NCCL runtime module so that the communicators are
non-blocking
and so that more consistent logic is used for handling errors. This
helps
resolve issues where the test may deadlock (either because of an
incorrect
runtime implementation issue or because of system config issue) without
errors being printed to stderr.

Additional TODOs are noted where the implementation can be further
improved.

GitOrigin-RevId: 5321de2a3d779500436c7a62097e0fc219958caf

Co-authored-by: Christopher Bate <cbate@nvidia.com>