Move internal changes (#380)
This PR moves the following internal changes to OSS
repo.

## [plan][transforms] Fix an issue in shape materialization pass 

This MR fixes an issue in `SimplifyExtractOfReshape` pattern
of shape materialization pass. This patten was being
applied even when reshape op operand is dynamic. However,
with dynamic operand, mapping extract index into
reshape operand doesn't work.
With this change, we return failure if reshape op
operand is dynamic. MLIR test is added for scenario
when pattern should return failure.

## [tensorrt] Fix incorrect handling of dynamic shape in
`tensorrt-broadcast-elimination`

Fixes an issue where a `tensorrt-broadcast-elimination` would improperly
handle
dynamically shaped tensors when attempting to reshape them. In certain
cases (when
more than 1 dynamic dimension is present), to perform a reshape, the
target shape
must be explicitly calculated in the IR and a dynamic reshape must be
created.

Co-authored-by: Copybara Bot <mlir-tensorrt-devs@exchange.nvidia.com>