[compiler][conversion] Fix an issue in `stablehlo.slice` conversion
This PR fixes an issue in `stablehlo.slice` conversion which
popped result type is dynamic.
In the converter, we use `tensorrt.slice` op builder that
infers result based on `static_size` (output shape is same as size on
TensorRT side) and `size` was taken as shape of stablehlo op output.
This caused two issues,
- `truncateI64ToI32` failed for dynamic output because dynamic dim can't be
truncated to i32 without loss.
- A new builder needed which could set result to dynamic
shape even when size is static.

With this change,
- `size` is computed as `ceil(limit-start)/stride` (these are
stablehlo slice op attributes) and NOT as shape of output.
- A new builder is added.

MLIR test is added.