Removes Parameter API
Removes the Parameter API which was used to denote constants in computation graphs.
We can instead use regular `Tensor`s since anything not specified as an input to the
compile call is treated as a constant.

`DefaultParameter` is preserved as an internal implementation detail for efficiently
defining expected parameter shapes/dtypes without materializing data.