Allows `DimensionSize`s to be used in `InputInfo`
Introduces a new `IntLike` type alias that covers integers and `DimensionSize`s
and updates `InputInfo` to accept shape arguments of this new type. Note that
the `DimensionSize`s must be evaluated in order to populate the profile correctly.

This allows for using the shapes of eager mode tensors when compiling.