[mlir-tensorrt] Add additional missing StableHLO patch
Added a patch that was missing from the last StableHLO upgrade. This patch
addresses some issues mentioned in https://github.com/openxla/stablehlo/issues/2634.

An additional test is added to mlir-tensorrt as a regression test.

Also makes some minor cleanup to CI/CD config in order to fix the caching
mechanism, using the PR as a test-case.