Disables pushing benchmark results for forks
It is not easily possible to share artifacts between workflows. Instead,
we will simply not upload benchmark results from a PR from a fork. The only
edge case here would be when such a PR introduces a new performance test. In
that case, the first benchmarks for that test would only be uploaded from the
first non-fork PR.