Reworks how constants are handled
The `Constant` op is used in two ways:

1. To represent constants in the network with `tensorrt.constant`.
    In this case, the data must be in CPU memory.

2. To represent inputs to compiled executables.
    In this case, the data must be in GPU memory and we only use `Constant`
    for its memref, i.e. we do not actually materialize MLIR operations with `to_mlir()`.

This commit updates `Constant.to_mlir()` to reject GPU memory since it
will only be called in case (1). Additionally, `Tensor.eval()` is updated
to evaluate Constants in CPU memory so that they are copied to GPU memory.

In some cases, we don't want to perform this copy, so an extra `_eval_for_internal_methods()`
is provided for internal use.