Adds support for source code inspection in guide examples, QDP tests
`triton` requires source code inspection in order to compile the kernel.
Since the generated docs dynamically execute code blocks, this is normally
not possible. This commit adds a special `ExecNamespace` that will inject
source code information as functions are defined.

This commit also adds various tests for QDPs and refactors the guide
and some of the code.