Speeds up Tensor.__init__ in certain cases, optimizes Constant.__init__
Speeds up Tensor.__init__ if the input data is a DLPack tensor and no cast/copy
is needed. This usage of the constructor is often seen in runtime code that
interoperates with frameworks, and so we want to minimze any overheads introduced
by Tripy.

This change disables the function registry dispatch logic for __init__ and
enables a fast path that skips fetching stack information. Also drops the explicit
`fast_init` constructor since we now automatically determine when the fast path can be used.

Also optimizes Constant.__init__ with a custom implementation that avoids
the overheads of the more general parent constructor.