Speed up GroupNorm, improve normalization testing (#637)
Due to an API incompatibility issue with Torch GroupNorm and TensorRT
GroupNorm, the implementation uses InstanceNorm instead as a workaround
(following the same WAR used by the ONNX parser and Torch -> ONNX
converter).

The latest opset of ONNX supports scale and bias with shape
`(num_channels,)`, so it is likely that the TRT API will see this
eventually supported, at which point we can switch to the most direct
implementation.

Running 1k iterations in a loop shows the new implementation is roughly
17% faster on average. The nsys trace shows that the new implementation
is significantly better fused (only two computation kernels, for the
instancenorm and the affine transform), and the performance for a single
iteration of the module is up to 40% faster (30µs vs 50µs).