Greatly limits how the dtype and device can be used in Tensor
Previously, setting `dtype` or `device` in the Tensor constructor
would incur a cast or copy operation respectively if the input data
were not already of the correct data type or on the right device.
These would be immediately evaluated, which could lead to extremely long tracing times.

Additionally, the implicit behavior is not in line with the rest of Tripy's
API design. With this change, users must explicitly call `cast` or `copy`.
This makes it clear *when* such an operation is actually happening.