[compiler] Bring `stablehlo-ext-constant-folding` up-to-par with `stablehlo-aggressive-folders`
This change updates patterns in `stablehlo-ext-constant-folding` to
ensure we can pass the upstream tests for
`stablehlo-aggressive-folders`.

A couple additional patterns were needed for simple binary elementwise
op folders.

Two bugs/deficiencies were found and corrected:

- Our utility for performing conversions between element types was not
  correctly returning failure during float->integer element
  conversions if a non-convertible value was present in the source
  array (e.g. NaN or Inf).

- Our pattern to fold `stablehlo.transpose` would crash on 0-d input
  tensors.

Note that we previously did not (and still do not) use the upstream
`stablehlo-aggressive-folders` patterns due to some deficiencies
related to enforcing the size limits. The upstream patterns don't
consistently enforce size limits on folder patterns when the
operand/result types are integer tensors. In addition, they often
abort in cases where the op result can be inferred to be static but
the specified type is a generalization of the inferred type. We handle
this through cast insertion. We should work on contributing upstream
to fix these deficiencies, which will allow us to depuplicate the
patterns in `stablehlo-ext-constant-folding` and
`stablehlo-aggressive-folders`.

GitOrigin-RevId: cf38cacfdc9c2665fb7298a757041be283fa4523