[compiler] Update CallOpInterface handling in TensorKindAnalysis
The MLIR backwards sparse dataflow framework has a hook called
`visitCallOperand`. Despite the name, its original purpose is to allow
analyses to specify how to handle CallOpInterface operands that are
*not* tied to the block arguments of the callee.

In the TensorKindAnalysis, we were assuming that this is the only
purpose of the hook and have been returning "host" for all such
operands. However, this is actually incorrect since by default the
`SparseBackwardDataFlowAnalysis` will also call `visitCallOperand` for
each operand when the callee is not visible to the solver (e.g. because
it is an SSA value only a declaration).

This commit updates the TensorKindAnalysis to explicitly implement
`visitExternalCall`, which gets us closer to the desired behavior. In
`visitExternalCall`, we set the lattice to "host" for all operands that
are not forwarded and invoke `setToExitState` for the operands that are
forwarded.

In the future we should consider constraints specified on the callee's
argument attributes, but there is the additional complication that we
don't know whether a concurrent pass is changing that IR, so it is
technically not valid to access.

We can perhaps can sidestep most issues by only running analysis in
dedicated module-scoped analysis passes that just cache the result for
the more fine-grained function passes to access.

GitOrigin-RevId: 61b1af910f50704b5b9c9cc55605481ea92fa617