[compiler] Add support for stablehlo.optimization_barrier
This change introduces `plan.optimization_barrier` which is a utility
operation that is used to prevent some optimizations which match
patterns in the input IR (e.g. constant folding) from being applied.
Semantically, the operation is an identity -- the operands are tied to
the results -- but it disrupts potential pattern matching.

Note that the operation is still functionally pure, so it does not
prevent optimizations like loop-invariant code motion or CSE.

We can use this as a conversion target for
`stablehlo.optimization_barrier` which has the equivalent semantic in
Stablehlo.

We only need `plan.optimization_barrier` until after bufferization. So
to eliminate `plan.optimization_barrier` from the IR, we simply add a
bufferization implementation that forwards the operands to the results.

GitOrigin-RevId: 6d724a00d58fd159a1845ceec8556e5c7bf35c1e