[compiler] Fix bufferization uses of `bufferization.materialize_in_destination`
In (previous commit) tried to fix a number of bufferization issues
dealing with our use of `bufferization.materialize_in_destination` and
`scf.if` ops. However, I discovered a better solution, which I implement
in this change.

The problem is that `bufferization.materialize_in_destination` is not
just a tensor-land copy operation. It is meant to indicate that the
target of the copy must be the buffer which will be associated with the
`dest` SSA, and it must be bufferized in-place. Bufferization will raise
an error if the bufferization does not occur in-place. This is useful
for indicating that e.g. the resulting bufferized IR *must* copy a
source data into a particular Value with important meaning (e.g.
function output argument).

However, we were using it in a couple places (namely convert all
`tensor.cast` to `tensor.empty` +
`bufferization.materialize_in_destination` ops) where the "in place"
requirement is not necessary. This was causing bufferization failures in
edge cases associated with `scf.if`.

To fix this, we just need an alternate copy-like operation that is
bufferizable and a DesitnationStyleOp. Luckily, there is already
`linalg.copy`, which we can use as a drop-in replacement. Then, to
recover the original behavior, we just convert the `linalg.copy` to
`memref.copy` operations.

GitOrigin-RevId: 93ed038f47690d33633db23be5e8f70b3d89d119