Release Builtin dialect's resource attribute blobs when top-level module is destroyed.
Adds a utility class that will release the blobs backing the builtin
dialect's top-level resource attributes when a module is destroyed. This
is then used in a downstream compiler (where we know this should be safe
since we hold a mutex during the compilation function and we know there
is no sharing of builtin ResourceElementsAttrs keys between compilation
of different modules).

GitOrigin-RevId: 7f308c7afdb139e69f1aba994888a04061b98b9d