[compiler] Fix end-to-end handling of `stablehlo.reverse`
Fixes end-to-end handling of `stablehlo.reverse`. The `stablehlo-legalize-to-linalg`
pass performs reversal at the indexing map level, generating indexing
maps like `(d0, d1) -> (d0, -d1 + [size-1])`. This is poorly supported
in the Linalg dialect, so we rewrite it to use explicit
`linalg.index/tensor.extract` in the body of the `linalg.generic` op.

GitOrigin-RevId: 2a5404936c2980ce75b3ad9dff560090212f2a73