Improve process for building Python wheels
Previously, we relied on custom CMake for generating Python wheels. We
assemble the python packages under `<build>/python_packages`. Afer the
build runs, all Python files and required binary objects are present
under that directory. Running a `ninja -C build
mlir-tensorrt-compiler-wheel` command then just invokes `pip wheel
<build>/python_packages/mlir_tensorrt_compiler` to generate the wheel
file.

While this process worked, it has one major deficiency which is that
typically binaries which are meant to be distributed are generated
by invoking a cmake "install" command (`cmake --install`). The CMake
install copies binaries to their final installed location and may perform
other post-processing steps such as stripping or changing various
ELF metadata.

In addition (and this is subject to opinion), Python users often expect
that a package can be wholly built from source just be running the
`pip install .` or `pip wheel .` command in the approriate directory.
This is a "Python first" approach to building release packages, and
actually it does somewhat simplify the process (from CI's point of view)
in terms of building the package under a large number of different
Python versions.

To address these issues, I have upgraded the `mlir-tensorrt-compiler`
package's `setup.py` to be able to just run the `pip wheel` command
(or `uv build`), and the Python build script will take care of invoking
CMake. It runs the CMake config, build, and install steps, then produces
the wheel file from the install tree (as opposed to the build tree).

Details are in the updated documentation.

GitOrigin-RevId: eb3cfd1979d520899d4381c0138584dacfb11050