[compiler] Don't offload slices of block arguments to tensorrt
We have encountered performance issues with offloading slices of block
arguments to TensorRT. This change disables offloading for these cases.

In addition, this change makes it more likely to encounter a warning
related to input alignment requirements in the runtime. Drop this
warning since it incorrect and no longer required by TensorRT.

GitOrigin-RevId: a03cc8404f373fb9816a5fc223fc61904ce69907