[tensorrt] Don't offload ops yielded from loops to TensorRT
If an elementwise operation or slice-update loperation is yielded from
a loop and one of the arguments is a block argument, then it's very
likely that the optimal solution is to bufferize that operation in-place
while re-using one of the input buffers for the output buffer. If such
an operation is offloaded to TensorRT, then such a bufferization would
be impossible due to I/O aliasing constraints. Therefore, we should
detect this situation and not offload such operations to TensorRT.

GitOrigin-RevId: e6169d5ffa91576f578d40834b2154dd138e88c7