Transpose Reshape Elimination pass.
This is a new pass that is designed to replace the existing Transpose
and Reshape Elemination passes.  This pass adds a lot of new rewrite
rules which enable pushing the transposes and reshapes around so that
they can be combined and then eliminated.

Signed-off-by: Matthew Francis-Landau <matthew.francis-landau@gm.com>

fix transpose lit test

Signed-off-by: Matthew Francis-Landau <matthew.francis-landau@gm.com>

cp

cp

cp

Signed-off-by: Matthew Francis-Landau <matthew.francis-landau@gm.com>

cp

cp

some improvements for matching matmul with reshapes, and update lit tests

Signed-off-by: Matthew Francis-Landau <matthew.francis-landau@gm.com>